requirements.txt
streamlit==1.19.0
tensorflow==2.18.0
numpy==1.24.0
Pillow==9.4.0
scikit-learn==1.2.0
